Corrected a bug reported by Lukjel: I found in DTD that GC element should have attributes: x,y,w,h. I can't see x and y in source... It looks, that there is a bug: x and y attributes will never appear.
Updated the DTD to be a bit:
- Reviced comments and given example
- Changed gc w and h attributes from #IMPLIED to more accurate "1"
- Changed gc x and y attributes from #IMPLIED to #REQUIRED
- Changed gridlayout w and h attributes from #IMPLIED to #REQUIRED

The last two changes are incompatible with previous DTD, but it can be considered to be fix for bug in the last revision. The comments in the DTD any way state that it should written as REQUIRED and IMPLIED attributes make no sense anyway.